+++ src/interactive/mod.rs
@@ -8,7 +8,7 @@ use ratatui::layout::{Alignment, Position, Rect};
 use ratatui::text::Span;
 use ratatui::widgets::Paragraph;
 use ratatui::{Frame, Terminal, crossterm};
-use rumqttc::{Client, Connection};
+use rumqttc::{Client, Connection, QoS};
 
 use self::ui::ElementInFocus;
 use crate::cli::Broker;
@@ -61,10 +61,11 @@ pub fn show(
     connection: Connection,
     broker: &Broker,
     subscribe_topic: Vec<String>,
+    qos: QoS,
     payload_size_limit: usize,
 ) -> anyhow::Result<()> {
     let mqtt_thread =
-        mqtt_thread::MqttThread::new(client, connection, subscribe_topic, payload_size_limit)?;
+        mqtt_thread::MqttThread::new(client, connection, subscribe_topic, qos, payload_size_limit)?;
     let app = App::new(broker, mqtt_thread);
 
     let original_hook = std::panic::take_hook();