+++ src/main.rs
@@ -2,7 +2,6 @@ use std::time::Duration;
 
 use clap::Parser as _;
 use cli::Subcommands;
-use rumqttc::QoS;
 
 mod clean_retained;
 mod cli;
@@ -22,12 +21,13 @@ fn main() -> anyhow::Result<()> {
     } else {
         None
     };
+    let qos = rumqttc::qos(matches.qos).unwrap();
     let (broker, client, connection) = mqtt::connect(matches.mqtt_connection, keep_alive)?;
 
     match matches.subcommands {
         Some(Subcommands::CleanRetained { topic, dry_run, .. }) => {
-            client.subscribe(topic, QoS::AtLeastOnce)?;
-            clean_retained::clean_retained(&client, connection, dry_run);
+            client.subscribe(topic, qos)?;
+            clean_retained::clean_retained(&client, connection, qos, dry_run);
         }
         Some(Subcommands::Log {
             topic,
@@ -35,7 +35,7 @@ fn main() -> anyhow::Result<()> {
             verbose,
         }) => {
             for topic in topic {
-                client.subscribe(topic, QoS::AtLeastOnce)?;
+                client.subscribe(topic, qos)?;
             }
             log::show(connection, json, verbose);
         }
@@ -45,7 +45,7 @@ fn main() -> anyhow::Result<()> {
             pretty,
         }) => {
             for topic in topic {
-                client.subscribe(topic, QoS::AtLeastOnce)?;
+                client.subscribe(topic, qos)?;
             }
             read_one::show(&client, connection, ignore_retained, pretty);
         }
@@ -66,7 +66,7 @@ fn main() -> anyhow::Result<()> {
                 },
                 String::into_bytes,
             );
-            client.publish(topic, QoS::AtLeastOnce, retain, payload)?;
+            client.publish(topic, qos, retain, payload)?;
             publish::eventloop(&client, connection, verbose);
         }
         None => {
@@ -75,6 +75,7 @@ fn main() -> anyhow::Result<()> {
                 connection,
                 &broker,
                 matches.topic,
+                qos,
                 matches.payload_size_limit,
             )?;
             client.disconnect()?;